requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
langchain==0.0.340
tiktoken==0.5.1
streamlit==1.28.2
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.2
numpy==1.26.2
torch==2.1.1
transformers==4.35.2
accelerate==0.25.0
bitsandbytes==0.41.1 